fix: missed one more...

+++ src/server/game/Entities/Player/Player.cpp
@@ -15981,7 +15981,7 @@ float Player::GetAverageItemLevelForDF()
     uint32 count = 0;
     uint8 level = GetLevel();
 
-    for (int i = EQUIPMENT_SLOT_START; i < EQUIPMENT_SLOT_END; ++i)
+    for (uint8 i = EQUIPMENT_SLOT_START; i < EQUIPMENT_SLOT_END; ++i)
     {
         // don't check tabard, ranged, offhand or shirt
         if (i == EQUIPMENT_SLOT_TABARD || i == EQUIPMENT_SLOT_RANGED || i == EQUIPMENT_SLOT_OFFHAND || i == EQUIPMENT_SLOT_BODY)
